--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/application/controller/CommandAchievementsController.java
@@ -1,12 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.application.controller;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("achievements")
-@Slf4j
-public class CommandAchievementsController {
-
-}
