--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/application/dto/AchievementsDTO.java
@@ -1,5 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.application.dto;
-
-public class AchievementsDTO {
-
-}
