--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/application/service/CommandAchievementsServiceImpl.java
@@ -1,8 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.application.service;
-
-
-import com.sixesSense.recorder.AchievementsDomain.command.domain.service.CommandAchievementsService;
-
-public class CommandAchievementsServiceImpl implements CommandAchievementsService {
-
-}
