--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/domain/service/CommandAchievementsService.java
@@ -1,5 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.domain.service;
-
-public interface CommandAchievementsService {
-
-}
