--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/ininfrastructure/repository/Repository.java
@@ -1,4 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.ininfrastructure.repository;
-
-public class Repository {
-}
