--- src/main/java/com/sixesSense/recorder/AchievementsDomain/command/ininfrastructure/service/InfraCommandAchievementsService.java
@@ -1,8 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.command.ininfrastructure.service;
-
-import org.springframework.stereotype.Service;
-
-
-public class InfraCommandAchievementsService {
-
-}
