--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/application/controller/QueryAchievementsController.java
@@ -1,9 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.application.controller;
-
-
-
-
-
-public class QueryAchievementsController {
-    
-}
