--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/application/service/QueryAchievementsServiceImpl.java
@@ -1,9 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.application.service;
-
-import com.sixesSense.recorder.AchievementsDomain.query.domain.service.QueryAchievementsService;
-
-
-
-public class QueryAchievementsServiceImpl implements QueryAchievementsService {
-
-}
