--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/domain/aggregate/dto/AchievementsDTO.java
@@ -1,4 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.domain.aggregate.dto;
-public class AchievementsDTO {
-
-}
