--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/domain/repository/AchievementsMapper.java
@@ -1,7 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.domain.repository;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface AchievementsMapper {
-}
