--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/domain/service/QueryAchievementsService.java
@@ -1,5 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.domain.service;
-
-public interface QueryAchievementsService {
-
-}
