--- src/main/java/com/sixesSense/recorder/AchievementsDomain/query/infrastructure/service/InfraQueryAchievementsService.java
@@ -1,7 +1,0 @@
-package com.sixesSense.recorder.AchievementsDomain.query.infrastructure.service;
-
-import org.springframework.stereotype.Service;
-
-
-public class InfraQueryAchievementsService {
-}
