+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/application/controller/CommandAchievementsController.java
@@ -1,0 +1,12 @@
+package com.sixesSense.recorder.achievementsDomain.command.application.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("achievements")
+@Slf4j
+public class CommandAchievementsController {
+
+}
