+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/application/dto/AchievementsDTO.java
@@ -1,0 +1,5 @@
+package com.sixesSense.recorder.achievementsDomain.command.application.dto;
+
+public class AchievementsDTO {
+
+}
