+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/application/service/CommandAchievementsServiceImpl.java
@@ -1,0 +1,8 @@
+package com.sixesSense.recorder.achievementsDomain.command.application.service;
+
+
+import com.sixesSense.recorder.achievementsDomain.command.domain.service.CommandAchievementsService;
+
+public class CommandAchievementsServiceImpl implements CommandAchievementsService {
+
+}
