+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/domain/repository/AchievementsRepository.java
@@ -1,0 +1,10 @@
+package com.sixesSense.recorder.achievementsDomain.command.domain.repository;
+
+import com.sixesSense.recorder.achievementsDomain.command.domain.aggregate.entity.Achievements;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface AchievementsRepository extends JpaRepository<Achievements, Integer> {
+}
