+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/domain/service/CommandAchievementsService.java
@@ -1,0 +1,5 @@
+package com.sixesSense.recorder.achievementsDomain.command.domain.service;
+
+public interface CommandAchievementsService {
+
+}
