+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/infrastructure/repository/Repository.java
@@ -1,0 +1,4 @@
+package com.sixesSense.recorder.achievementsDomain.command.infrastructure.repository;
+
+public class Repository {
+}
