+++ src/main/java/com/sixesSense/recorder/achievementsDomain/command/infrastructure/service/InfraCommandAchievementsService.java
@@ -1,0 +1,6 @@
+package com.sixesSense.recorder.achievementsDomain.command.infrastructure.service;
+
+
+public class InfraCommandAchievementsService {
+
+}
