+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/application/controller/QueryAchievementsController.java
@@ -1,0 +1,9 @@
+package com.sixesSense.recorder.achievementsDomain.query.application.controller;
+
+
+
+
+
+public class QueryAchievementsController {
+    
+}
