+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/application/service/QueryAchievementsServiceImpl.java
@@ -1,0 +1,9 @@
+package com.sixesSense.recorder.achievementsDomain.query.application.service;
+
+import com.sixesSense.recorder.achievementsDomain.query.domain.service.QueryAchievementsService;
+
+
+
+public class QueryAchievementsServiceImpl implements QueryAchievementsService {
+
+}
