+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/domain/aggregate/dto/AchievementsDTO.java
@@ -1,0 +1,4 @@
+package com.sixesSense.recorder.achievementsDomain.query.domain.aggregate.dto;
+public class AchievementsDTO {
+
+}
