+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/domain/repository/AchievementsMapper.java
@@ -1,0 +1,7 @@
+package com.sixesSense.recorder.achievementsDomain.query.domain.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface AchievementsMapper {
+}
