+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/domain/service/QueryAchievementsService.java
@@ -1,0 +1,5 @@
+package com.sixesSense.recorder.achievementsDomain.query.domain.service;
+
+public interface QueryAchievementsService {
+
+}
