+++ src/main/java/com/sixesSense/recorder/achievementsDomain/query/infrastructure/service/InfraQueryAchievementsService.java
@@ -1,0 +1,5 @@
+package com.sixesSense.recorder.achievementsDomain.query.infrastructure.service;
+
+
+public class InfraQueryAchievementsService {
+}
