+++ src/main/resources/mappers/AchievementsMapper.xml
@@ -2,6 +2,6 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.sixesSense.recorder.AchievementsDomain.query.domain.repository.AchievementsMapper">
+<mapper namespace="com.sixesSense.recorder.achievementsDomain.query.domain.repository.AchievementsMapper">
 
 </mapper>
