+++ src/test/java/com/sixesSense/recorder/RecoderApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class RecoderApplicationTests {
+class RecorderApplicationTests {
 
     @Test
     void contextLoads() {
